+++ include/carbon/compressor/carbon-compressor-huffman.h
@@ -40,6 +40,9 @@ carbon_compressor_huffman_write_extra(carbon_compressor_t *self, carbon_memfile_
                                       const carbon_vec_t ofType (const char *) *strings);
 
 CARBON_EXPORT(bool)
+carbon_compressor_huffman_read_extra(carbon_compressor_t *self, FILE *src, size_t nbytes);
+
+CARBON_EXPORT(bool)
 carbon_compressor_huffman_print_extra(carbon_compressor_t *self, FILE *file, carbon_memfile_t *src);
 
 CARBON_EXPORT(bool)