+++ src/carbon/compressor/carbon-compressor-none.c
@@ -61,6 +61,18 @@ carbon_compressor_none_write_extra(carbon_compressor_t *self, carbon_memfile_t *
     return true;
 }
 
+CARBON_EXPORT(bool)
+carbon_compressor_none_read_extra(carbon_compressor_t *self, FILE *src, size_t nbytes)
+{
+    CARBON_CHECK_TAG(self->tag, CARBON_COMPRESSOR_NONE);
+
+    CARBON_UNUSED(self);
+    CARBON_UNUSED(src);
+    CARBON_UNUSED(nbytes);
+    /* nothing to do for uncompressed dictionaries */
+    return true;
+}
+
 bool carbon_compressor_none_print_extra(carbon_compressor_t *self, FILE *file, carbon_memfile_t *src)
 {
     CARBON_CHECK_TAG(self->tag, CARBON_COMPRESSOR_NONE);